feat(module:table): split th component
