+++ components/table/style/patch.less
@@ -1,7 +1,10 @@
 nz-table,
 nz-table-title-footer,
 nz-table-inner-scroll,
-nz-table-inner-default {
+nz-table-inner-default,
+nz-table-selection,
+nz-table-sorters,
+nz-table-filter {
   display: block;
 }
