+++ components/table/table.types.ts
@@ -19,3 +19,10 @@ export type NzTableLayoutType = 'fixed' | 'auto';
 export type NzTablePaginationPositionType = 'top' | 'bottom' | 'both';
 
 export type NzTableSizeType = 'middle' | 'default' | 'small';
+
+export type NzThFilterType = Array<{ text: string; value: NzSafeAny; byDefault?: boolean }>;
+export interface NzThItemInterface {
+  text: string;
+  value: NzSafeAny;
+  checked: boolean;
+}
